Rework API page iteration, dog_json2csv.py
Fixing a bug where a request with one page caused every item to be saved to the output twice, the first time at the start of the while loop and again when the next page api_url page=None got the information for page one again.

Now the first page is saved to the list and if there is more than one page the script iterates over page 2 - page end to save them to the list.